SettingsScreen structure + interactor implementations

+++ app/src/main/java/szarch/bme/hu/ibdb/network/models/review/ReviewRequest.kt
@@ -2,5 +2,5 @@ package szarch.bme.hu.ibdb.network.models.review
 
 data class ReviewRequest(
     val comment: String?,
-    val points: Int
+    val points: Double
 )
